requirement.txt
streamlit==1.34.0
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.10.0
pinecone-client==3.0.2
openai==1.30.1
python-dotenv==1.0.1
selenium==4.20.0
webdriver-manager==4.0.1
scipy==1.13.1
numpy==1.26.4
av==12.0.0
streamlit-webrtc==0.47.0
tqdm==4.66.4